refactor: types 별도 파일로 분리 (#178)
